Dev.Settings: Add option to control whether to backup/restore hosts file
By default, LineageOS always backs up the /system/etc/hosts file and
restores it during an upgrade. Reason is, that the pre-delivered hosts
file by LineageOS only contains the loopback definition. If a root app to
edit the hosts file is used, or the user has edited his hosts file via
adb root, the hosts file should persist.

This ROM version however delivers a pre-defined large hosts file with many
blocked ad- and malware sites, which is regularly updated. Therefore, this
behavior is indeed not desired.

The changed default behavior therefore is NOT any longer to automatically
backup/restore the hosts file. A new switch in the Development options
will however allow the user to choose to keep the previous LineageOS
default behavior. This switch is only visible for the admin-user '0'.

If activated, file /data/user_de/0/com.android.settings/files/.hosts.backup
is created; if deactivated, this empty file is purged.
The backup/restore is done by the installer shell script 50-lineage.sh
The behavior of this script is altered accordingly in a separate commit to
check for the existence of this file.

Cherry-picked from lin17-microG repo

Change-Id: Icf9c225d6568173ce0ddd15311ed2f49d0a0132b